refactor: move cart logic to separate hook
