+++ src/app.tsx
@@ -13,6 +13,7 @@ import { LoginPage } from './pages/login/login';
 import { ProductsProvider } from './components/hooks/useProducts';
 import { ProductPage } from './pages/product-page/product-page';
 import { UserPage } from './pages/user/UserPage';
+import { CartProvider } from './components/hooks/useCart';
 
 export function App() {
   return (
@@ -21,21 +22,23 @@ export function App() {
         <BrowserRouter>
           <AuthProvider>
             <ProductsProvider>
-              <Routes>
-                <Route element={<MainLayout />}>
-                  <Route path="/" element={<HomePage />} />
-                  <Route path="/catalog" element={<CatalogPage />} />
-                  <Route path="/catalog/:categoryKey" element={<CatalogPage />} />
-                  <Route path="/catalog/:categoryKey/:subcategoryKey" element={<CatalogPage />} />
-                  <Route path="/products/:productId" element={<ProductPage />} />
-                  <Route path="/about-us" element={<AboutPage />} />
-                  <Route path="/login" element={<LoginPage />} />
-                  <Route path="/registration" element={<RegistrationPage />} />
-                  <Route path="/404" element={<NotFoundPage />} />
-                  <Route path="*" element={<NotFoundPage />} />
-                  <Route path="/user" element={<UserPage />} />
-                </Route>
-              </Routes>
+              <CartProvider>
+                <Routes>
+                  <Route element={<MainLayout />}>
+                    <Route path="/" element={<HomePage />} />
+                    <Route path="/catalog" element={<CatalogPage />} />
+                    <Route path="/catalog/:categoryKey" element={<CatalogPage />} />
+                    <Route path="/catalog/:categoryKey/:subcategoryKey" element={<CatalogPage />} />
+                    <Route path="/products/:productId" element={<ProductPage />} />
+                    <Route path="/about-us" element={<AboutPage />} />
+                    <Route path="/login" element={<LoginPage />} />
+                    <Route path="/registration" element={<RegistrationPage />} />
+                    <Route path="/404" element={<NotFoundPage />} />
+                    <Route path="*" element={<NotFoundPage />} />
+                    <Route path="/user" element={<UserPage />} />
+                  </Route>
+                </Routes>
+              </CartProvider>
             </ProductsProvider>
           </AuthProvider>
         </BrowserRouter>
