+++ src/pages/product-page/product-page.tsx
@@ -1,12 +1,12 @@
 import { useParams, useNavigate } from 'react-router-dom';
 import { useProducts } from '../../components/hooks/useProducts';
+import { useCart } from '../../components/hooks/useCart';
 import { Card, Text, Spin, Button, useToaster } from '@gravity-ui/uikit';
 import { useEffect, useState } from 'react';
 import { NotFoundPage } from '../404/not-found';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import { Navigation, Pagination } from 'swiper/modules';
 import { ChevronLeft, ChevronRight, Xmark } from '@gravity-ui/icons';
-import { customerAPI } from '../../api/customer-api';
 import 'swiper/css';
 import 'swiper/css/navigation';
 import 'swiper/css/pagination';
@@ -22,23 +22,13 @@ type AttributeValue = AttributeValueObject | AttributeValueObject[] | string | n
 
 export function ProductPage() {
   const { productId } = useParams();
-  const { productDetails, getProductDetails, isLoading, error, notFound, addToCart, isProductInCart, removeFromCart } =
-    useProducts();
+  const { productDetails, getProductDetails, isLoading, error, notFound } = useProducts();
+  const { addToCart, isProductInCart, removeFromCart } = useCart();
   const [initialSlide, setInitialSlide] = useState(0);
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [isInCart, setIsInCart] = useState(false);
   const navigate = useNavigate();
   const toaster = useToaster();
-
-  const getFullCartInfo = async () => {
-    try {
-      const response = await customerAPI.apiRoot().me().carts().get().execute();
-      const cart = response.body.results[0];
-      console.log('Current cart info:', cart);
-    } catch (error) {
-      console.error('Error fetching full cart info:', error);
-    }
-  };
 
   const getLabel = (item: AttributeValue | AttributeValueObject): string => {
     if (typeof item === 'object') {
@@ -82,7 +72,6 @@ export function ProductPage() {
           content: 'Product added to cart!',
           theme: 'success',
         });
-        await getFullCartInfo();
       } catch (_error) {
         toaster.add({
           name: 'cart-error',
@@ -103,7 +92,6 @@ export function ProductPage() {
           content: 'Product removed from cart!',
           theme: 'info',
         });
-        await getFullCartInfo();
       } catch (_error) {
         toaster.add({
           name: 'cart-remove-error',
@@ -127,7 +115,6 @@ export function ProductPage() {
   useEffect(() => {
     if (productId) {
       getProductDetails(productId);
-      void getFullCartInfo();
     }
   }, [productId, getProductDetails]);
